ftx utils refactor
