--- src/exchanges/ftx/utils/functions/validatation.js
@@ -1,1 +1,0 @@
-module.exports = {};
