+++ src/exchanges/ftx/wallets/spot.js
@@ -1,7 +1,7 @@
 const FtxBase = require('../base');
 const mergeObjects = require('deepmerge');
 
-const utils = require('../utils');
+const utils = require('../../../base/utils/utils');
 
 const AccountDataStream = require('../streams/accountDataStream');
 const MarketDataStream = require('../streams/marketDataStream');
@@ -37,9 +37,14 @@ class FtxSpot extends FtxBase {
 
         // TODO: Cache update here
 
-        const spotMarkets = utils.extractSpotMarkets(markets.result);
+        const spotMarkets = markets.result.filter((market) => market.type === 'spot');
 
-        const tickers = utils.extractTickersFromMarkets(spotMarkets);
+        const tickers = spotMarkets
+            .map((market) => {
+                return [market.baseCurrency, market.quoteCurrency];
+            })
+            .flat()
+            .filter(Boolean);
 
         utils.linkOriginalPayload(tickers, markets);
 
@@ -54,7 +59,7 @@ class FtxSpot extends FtxBase {
 
         // TODO: Cache update here
 
-        const spotMarkets = utils.extractSpotMarkets(response.result);
+        const spotMarkets = response.result.filter((market) => market.type === 'spot');
 
         const markets = spotMarkets.map((m) => {
             return {
@@ -173,14 +178,14 @@ class FtxSpot extends FtxBase {
     async createOrder(zOrder) {
         this.validateOrderParams(zOrder);
 
-        const fOrder = utils.transfromZenfuseOrder(zOrder);
+        const fOrder = this.transformZenfuseOrder(zOrder);
 
         const fCreatedOrder = await this.privateFetch('api/orders', {
             method: 'POST',
             json: fOrder,
         });
 
-        const zCreatedOrder = utils.transfromFtxOrder(fCreatedOrder.result);
+        const zCreatedOrder = this.transformFtxOrder(fCreatedOrder.result);
 
         this.cache.cacheOrder(zCreatedOrder);
 
@@ -261,7 +266,79 @@ class FtxSpot extends FtxBase {
     async fetchOrderById(orderId) {
         const responce = await this.privateFetch(`api/orders/${orderId}`);
 
-        const zOrder = utils.transfromFtxOrder(responce.result);
+        const zOrder = this.transformFtxOrder(responce.result);
+
+        return zOrder;
+    }
+
+    /**
+     * @typedef {import('../../../../base/schemas/orderParams').ZenfuseOrderParams} OrderParams
+     */
+
+    /**
+     * Zenfuse -> FTX
+     *
+     * @param {OrderParams} zOrder Order from
+     * @returns {object} Order for ftx api
+     */
+    transformZenfuseOrder(zOrder) {
+        const TRANSFORM_LIST = ['side', 'type', 'price', 'quantity', 'symbol'];
+
+        const fOrder = {
+            market: zOrder.symbol,
+            type: zOrder.type,
+            side: zOrder.side,
+            size: zOrder.quantity,
+        };
+
+        if (zOrder.price) {
+            fOrder.price = zOrder.price;
+        }
+
+        if (zOrder.type === 'market') {
+            fOrder.price = null;
+        }
+
+        // Allow user extra keys
+        for (const [key, value] of Object.entries(zOrder)) {
+            if (!TRANSFORM_LIST.includes(key)) {
+                fOrder[key] = value;
+            }
+        }
+
+        return fOrder;
+    }
+
+    /**
+     * @typedef {import('../../../../base/schemas/openOrder').PlacedOrder} PlacedOrder
+     */
+
+    /**
+     * FTX -> Zenfuse
+     *
+     * @param {*} fOrder Order from FTX
+     * @returns {PlacedOrder} Zenfuse Order
+     */
+    transformFtxOrder(fOrder) {
+        /**
+         * @type {PlacedOrder}
+         */
+        const zOrder = {};
+
+        zOrder.id = fOrder.id.toString();
+        zOrder.timestamp = Date.parse(fOrder.createdAt);
+        zOrder.symbol = fOrder.market;
+        zOrder.type = fOrder.type;
+        zOrder.side = fOrder.side;
+        zOrder.quantity = parseFloat(fOrder.size);
+        zOrder.price = fOrder.price ? parseFloat(fOrder.price) : undefined;
+        // zOrder.trades = bOrder.fills; // TODO: Fill commision counter
+
+        if (fOrder.status === 'new') {
+            zOrder.status = 'open';
+        } else {
+            zOrder.status = fOrder.status;
+        }
 
         return zOrder;
     }
